Form for deleting

+++ delete.php
@@ -4,20 +4,25 @@ include('session.php');
 if($start == 1){
 
     $id=$_GET['id'];
-    $action=$_GET['action'];
 
     echo "<h3>Are you sure you want to remove this product?</h3>";
-    echo '<a href="?id='. $id .'&&action=yes">Yes</a>';
+    echo '<a href="?id='. $id .'&action=yes">Yes</a>';
     echo "<br>";
-    echo '<a href="?id='. $id .'&&action=no">No</a>';
+    echo "<br>";
+    echo '<a href="?id='. $id .'&action=no">No</a>';
 
-    if($action='yes'){
-        mysqli_query($conn,"DELETE FROM `products` WHERE id = '$id'");
-        header('location:db.php');
-    }elseif($action='no'){
-        header('location:db.php');
+
+    if(isset($_GET['action'])){
+
+        $action=$_GET['action'];
+
+        if($action='yes'){
+            mysqli_query($conn,"DELETE FROM `products` WHERE id = '$id'");
+            header('Location: db.php');
+        }elseif($action='no'){
+            header('Location: db.php');
+        }
     }
-
 
 }
 ?>
